refactor Matcher into it's own file
