+++ core/matchers/index.ts
@@ -1,9 +1,11 @@
 import { FunctionSpyCallCountMatcher } from "./function-spy-call-count-matcher";
 import { FunctionSpyMatcher } from "./function-spy-matcher";
+import { Matcher } from "./matcher";
 import { SpyCallCountType } from "./spy-call-count-type";
 
 export {
    FunctionSpyCallCountMatcher,
    FunctionSpyMatcher,
+   Matcher,
    SpyCallCountType
 }
